React && Chakra Setup (Part II)
Set a dark mode using Chakra

Create a styles object with the global function and changing the background and color in the body field. Light mode uses gray.100 for the bg. Dark mode uses #000 for the bg.

Pass styles into extendTheme so it will override the color that ChakraProvider gives us and then passing it into the App