Design Auth Page (Part II)
AuthForm returns a <> containing a Box containing a VStack containing an Image, 2 <Input/> and an <Input/> if isLogin.

VStack also contains Button which is Login if isLogin else is Sign Up.

VStack also contains a Flex containing a Box, Text and Box for the --- OR --- design.

Below the --- OR --- design is a Flex containing an Image and a Text for Log in with Google.