Design Sidebar (Part III)
Logout button is a Tooltip containing a Link containing a BiLogOut icon and Box containing the text Logout.

Tooltip label is "Logout", Link to is "/auth" and marginTop={"auto"} brings it to the bottom of the page.

Tooltip display={{base:"block", md:"none"}}, Box containing the text Logout display={{base:"none", md:"block"}}.