Design Home Page (Part II)
Create PostFooter.jsx, PostHeader.jsx

FeedPost.jsx returns a fragment containing a <PostHeader /> followed by a <Box> containing an <Image /> followed by a <PostFooter />

PostHeader.jsx returns a Flex with justifyContent={"space-between"}. This Flex contains a Flex and a Box.

The Flex contains an Avatar and a Flex containing text and a Box with more text.

The Box contains Text.