Design Home Page (Part III)
overflow={"hidden"} clips the Image to fit within the rounded corners of the Box

PostFooter.jsx returns a Box containing a Flex (like and comment logo), Text (number of likes), Text (username and caption), Text (viewing comments), Flex (adding a comment).

The useState hook manages the states liked and likes. The handleLike function updates the liked state and likes count.

The Flex contains a Box containing the NotificationsLogo or UnlikeLogo based on the liked state and handles the click event to like or unlike a post.

The Text uses {" "} for spacing and contains another Text as={"span"}

The Flex contains an InputGroup containing an Input and InputRightElement with a Button.

variant={"flushed"} gives the Input a bottom border which becomes visible when the input is focused.