Update AuthForm.jsx (Part II)
Signup.jsx uses two state variables: inputs and showPassword.

inputs is an object holding the form input values. showPassword is a Boolean that toggles the visibility of the password input.

Signup.jsx returns 3 Input components and one InputGroup.

The InputGroup contains an Input and an InputRightElement.

The InputRightElement contains a Button.