Update AuthForm.jsx (Part III)
Added a Sign Up button in Signup.jsx.

Extracted the email, password, and confirm password input forms into separate Login and Signup components in AuthForm.jsx.

Introduced a GoogleAuth component.

The size of inputs are set to sm in Login.jsx.

Added a Flex container with an image and text to allow users to log in with Google.