Signup With Email && Password (Part II)
Signup.jsx initializes the custom hook, extracting the loading, error and signup values from it.

The onClick handler to the Sign Up button calls the signup function, passing the inputs state as an argument.

The signup function checks if all fields are filled. If not, it logs a message and exits.

It then attempts to create a new user using the createUserWithEmailAndPassword function.

If user creation is successful, it creates a user document with fields uid, email, username, fullName, bio, profilePicURL, followers, following, posts, createdAt and stores it in Firestore.

If there is an error at any step, it logs the error.

The hook returns the loading and error states, as well as the signup function.