Logout User
Create useLogout.js, rafce

Sidebar.jsx uses the useLogout hook to get the handleLogout function and the isLoggingOut state.

The Link is replaced by Flex with an onClick handler that triggers the handleLogout function.

A Button shows the loadingState.

useLogout.js is used to get the signOut function, a loading state isLoggingOut and any error that occurs during the sign-out process.

The showToast function is obtained from useShowToast.

handleLogout is an async function that tries to sign out the user.

If successful, it removes the user-info from localStorage and logs a message to the console.

If an error occurs, it displays an error toast with the error message.

The hook returns the handleLogout function, the isLoggingOut state and any error.