Create authStore With Zustand (Part I)
Create store folder containing authStore.js

authStore.js defines a custom hook using the zustand library, which is a small, fast and scalable state management solution for React.

The create function creates a new state management store.

The useAuthStore constant is assigned the value returned by create.

create takes a function that returns an object.

This object defines the initial state of the store and the actions that can be performed.

The user property is set to null.

The login action takes a user as an argument and updates the state using the set function.

The logout action sets the user property to null.

The setUser action is similar to login.

useAuthStore is exported.