Create authStore With Zustand (Part II)
useSignUpWithEmailAndPassword retrieved the loginUser function from the useAuthStore hook.

loginUser is called with the userDoc to update the global state with the new user.

App.jsx added authUser which retrieves the current user from the global state.

If authUser exists, "/" renders the HomePage, otherwise it redirects to the AuthPage.

The "/auth" route renders the AuthPage only if there is no authUser, otherwise it redirects to "/".

authStore.js initialized user with localStorage value.