Create Navbar (Part I)
Create Navbar folder containing Navbar.jsx, rafce.

PageLayout.jsx conditionally renders a navbar when the user is not authenticated, the page is not /auth and loading is complete.

It uses a column layout when the navbar is rendered and a row layout otherwise.

Navbar.jsx is a Container containing a Flex containing an Image (Instagram logo) and a Flex (Login and Signup buttons).

The Flex contains 2 Link components each containing a Button.