Login User (Part I)
Create useLogin.js, rafce, delete return.

useLogin initializes the custom hook for showing toast notifications.

Destructures the signInWithEmailAndPassword function and the loading and error states from the Firebase hook.

Retrieves the login function from the authentication store to update the user's authentication state in the application.

login function checks if the email and password fields are filled.

Uses signInWithEmailAndPassword to authenticate the user with Firebase.

If authentication is successful, retrieve the user document from Firestore using the user's UID.

Stores the user data in localStorage and updates the application state with the user data.

Shows an error toast notification if something goes wrong during the login process.

Returns the loading and error states and the login function.