Login User (Part II)
Login.jsx uses the useLogin hook, which provides loading, error and login states/functions.

Conditional rendering block to display an error message if error exists.

The Button calls the login function from the useLogin hook when clicked, passing the inputs state as an argument.

Shows a loading spinner when the loading state is true, using the isLoading prop.