Login User (Part III)
The Avatar uses authUser.profilePicURL from the useAuthStore hook.

The Text displays authUser.username.

The profile picture and username are wrapped in Link components that navigate to the user's profile page using their username.

The Button is used instead of a Link.

The handleLogout function from the useLogout hook is used to handle the logout process.

The isLoggingOut state from the useLogout hook indicates if the logout process is ongoing, which is used to show a loading state for the button.

The name prop in Avatar is omitted.

The style={{textDecoration:"none"}} in Button is omitted.