Google Authentication (Part II)
GoogleAuth.jsx:
handleGoogleAuth is an async function that encapsulates the Google authentication and user data handling.
Is triggered when the component is clicked.
Handles errors and displays them using showToast.
Creates a new user document in Firestore if authentication is successful.
Stores user information in localStorage and updates the authentication store with loginUser.

SuggestedHeader.jsx
If there is no authenticated user, the component does not render.