Fetching Profile Header Data (Part III)
useShowToast.js:
The useCallback hook is used to memorize the function.
This prevents unnecessary recreations of the function on every render, which helps avoid infinite loops.

useGetUserProfileByUsername.js:
Returns an object containing isLoading and userProfile.

ProfilePage.jsx:
Uses the useGetUserProfileByUsername hook to fetch the user profile data based on the username from the URL parameters.