Fetching Profile Header Data (Part VI)
authUser gets the authenticated user from the useAuthStore.

visitingOwnProfileAndAuth checks if the authenticated user exists and if their username matches the username of the profile being viewed.

visitingAnotherProfileAndAuth checks if the authenticated user exists and that their username does not match the username of the profile being viewed.

Conditionally renders either the "Edit Profile" button or the "Follow" button.