Edit User Profile (Part I)
Create EditProfile.jsx which is a Modal containing a ModalOverlay and ModalContent.
The isOpen prop determines whether the modal is open  or not.
The onClose prop is a function that closes the modal.
boxShadow={"xl"} applies a shadow to the ModalContent.
lineHeight={1.1} specifies the amount of space between lines of text.
1.1 means that the space between lines is 1.1 times the size of the font.
direction={["column", "row"]} means that on smaller screens, items will be stacked vertically.
On larger screens, items will be stacked horizontally.
The Center component centers its children both horizontally and vertically.

ProfileHeader.jsx uses the useDisclosure hook from Chakra UI to manage the state of the modal (open/close).
The Edit Profile button now has an onClick event that triggers the onOpen function to open the modal.
The EditProfile component is conditionally rendered based on the isOpen state.