Edit User Profile (Part V)
Check if the selected file is an image and if the file size exceeds a maximum limit of 2MB.
Else an error toast is shown and selectedFile is set to null.
A FileReader instance is created to read the file as a data URL.
The reader.onloadend event handler sets the selectedFile state with the base64 string of the image when the file reading is complete.
The usePreviewImg hook returns an object that allows access to the selected image, handle file changes, and reset the selected file.