Follow && Unfollow (Part II)
The handleFollowUser function sets isUpdating to true to indicate that an update is in progress.

It then tries to update the Firestore database.

References to the current user's document and the user to be followed/unfollowed are created.

The updateDoc function is used to update the "following" field of the current user.

If the user is following, remove the user ID from the following array.

If the user is not following, add it.