Follow && Unfollow (Part V) useFollowUser.js: The updateDoc call modifies the followers field of the user being followed or unfollowed.
ProfileHeader.jsx adds optional chaining to ensure if userProfile is null or undefined, it won't cause an error.
Adds the onClick event handler to the Follow button, linking it to the handleFollowUser function.
Adds the isLoading prop to show a loading state when isUpdating is true.
The button text dynamically changes based on whether the user is following or not.