Search For User Profiles (Part II)
useSearchUser.js created a Firestore query to search for users where the username matches the input username.
Used getDocs to execute the query and fetch results.
If user not found, show an error toast.
If users are found, iterate through the querySnapshot and set the user state to the fetched user data.
Returned isLoading, getUserProfile and user, making these values accessible when the hook is used.

Search.jsx manages the opening and closing states for a modal with the useDisclosure hook.