Create Post (Part II)
The useState hook creates a caption state variable to manage the text entered in the Textarea component.
The useRef hook creates an imageRef for the file input element.
The usePreviewImg hook manages image selection and preview functionality.
It provides handleImageChange, selectedFile for storing the selected file URL and setSelectedFile.
Textarea now has value and onChange properties to bind it to the caption state variable, allowing the caption text to be updated as the user types.
Input is updated with a ref attribute and an onChange event handler to process file selection.
The BsFillImageFill icon includes an onClick event that triggers a click on the hidden file input.
If selectedFile is not empty, a Flex containing an Image and CloseButton is rendered.
Clicking the CloseButton will clear the selectedFile state.