Create Comment (Part II)
usePostComment.js:
addComment is a function extracted from usePostStore to update the local state after a comment is added.
Update the Firestore document of the post by appending the new comment to the comments array.
Calls addComment, showToast, return isCommenting and handlePostComment.

postStore.js:
The addComment function takes a postId and a comment and updates the relevant post in the state by appending the new comment to the comments array of the post.
Uses set to update the state, mapping over the posts array to find the correct post by postId and updating its comments array.