Get Details of The Comment (Part II)
Create utils folder containing timeAgo.js.

Comment.jsx:
When the data is still being fetched, it returns a CommentSkeleton.
Added a Link component around the Avatar and username to navigate to the user's profile page when clicked.
Used userProfile.profilePicURL for the Avatar source and userProfile.username for displaying the username.
Added a timeAgo function to show how long ago the comment was created.
The CommentSkeleton is a Flex containing a SkeletonCircle and Flex.