Fetch Feed Posts (Part VI)
FeedPosts.jsx:
Two Text components are added to handle the case when posts.length is 0 and isLoading is false to provide a user-friendly
message indicating that no posts are available.

FeedPost.jsx uncomments the PostFooter component and passes post and creatorProfile as props.

PostFooter.js uses creatorProfile instead of username.
The text display for the username and caption changes.
The comments section is conditionally rendered, displaying the number of comments if any exist.

PostFooter.jsx replaces "username" with creatorProfile.